fix bug in envelope, needed an epsilon(?) value as we would not always be exactly == 0. change pwm clock to target a sample rate, this  increases the bit depth to 13 bits which sounds nicer

+++ src/synth/include/envelope.h
@@ -4,7 +4,7 @@
 #include "audiocontext.h"
 #include "fixedpoint.h"
 
-void synth_envelope_init(uint32_t sampleRate);
+void synth_envelope_init(uint16_t sampleRate);
 
 fix16 synth_envelope_process(AudioContext_t* context);
